fixed #1439:synchronize to cocos2d-iphone v2.0.0

+++ cocos2dx/textures/CCTexturePVR.cpp
@@ -267,7 +267,7 @@ bool CCTexturePVR::unpackPVRData(unsigned char* data, unsigned int len)
                 }
 
                 dataSize = widthBlocks * heightBlocks * ((blockSize  * bpp) / 8);
-                float packetLength = (dataLength - dataOffset);
+                unsigned int packetLength = (dataLength - dataOffset);
                 packetLength = packetLength > dataSize ? dataSize : packetLength;
                 
                 //Make record to the mipmaps array and increment coutner
